Move federatedtypeconfig webhook to new package
